+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to Semantic Versioning.
 * File drag and drop events can now be detected via `Event::FileDropped`.
 * The clipboard can now be manipulated via `input::get_clipboard_text` and `input::set_clipboard_text`.
 * `input::get_key_modifier_down` and `input::get_key_modifier_up` have been added, allowing for code handling the control, alt and shift keys to be made more compact.
+* An `Animation` can now be set to stop playing after all the frames have been displayed, instead of looping. This can either be controlled by the `set_repeating` method, or you can create a non-looping animation directly by calling `Animation::once`.
 
 ## [0.3.4] - 2020-04-12